+++ rollkit/main.go
@@ -34,6 +34,8 @@ func (app) Execute(ctx context.Context, c *plugin.ExecutedCommand, _ plugin.Clie
 		return cmd.AddHandler(ctx, c)
 	case "init":
 		return cmd.InitHandler(ctx, c)
+	case "edit-genesis":
+		return cmd.EditGenesisHandler(ctx, c)
 	default:
 		return errors.Errorf("unknown command: %s", strings.Join(c.OsArgs, " "))
 	}