+++ rollkit/template/constants.go
@@ -10,8 +10,11 @@ const (
 
 const (
 	GoExecPackage = "github.com/rollkit/go-execution-abci"
-	GoExecVersion = "v0.2.1-0.20250625133753-4c5a41d10330" // TODO(@julienrbrt): use tag when available
+	GoExecVersion = "v0.3.0"
 
 	RollkitPackage = "github.com/rollkit/rollkit"
-	RollkitVersion = "v0.14.2-0.20250625130707-6ad581a97a16" // TODO(@julienrbrt): use tag when available
+	RollkitVersion = "v1.0.0-beta.2"
+
+	RollkitDaCmd     = "github.com/rollkit/rollkit/da/cmd/local-da"
+	RollkitDaVersion = "v1.0.0-beta.1"
 )